Add test to stat tracker for highest total score method
