+++ test/stat_tracker_test.rb
@@ -14,18 +14,6 @@ class TestStatTracker < Minitest::Test
     }
     @stat_tracker = StatTracker.from_csv(@locations)
   end
-  #
-  # def test_stat_tracker_can_pull_file_locations
-  #   assert_equal 7_441, @stat_tracker.games.length
-  #   assert_equal 32, @stat_tracker.teams.length
-  #   assert_equal 14_882, @stat_tracker.game_teams.length
-  # end
-  #
-  # def test_it_has_data
-  #   assert_equal '20122013', @stat_tracker.games[5].season
-  #   assert_equal 'Columbus Crew SC', @stat_tracker.teams[31].team_name
-  #   assert_equal 41.5, @stat_tracker.game_teams[14_028].faceOffWinPercentage
-  # end
 
   def test_it_can_get_team_data
     locations =  {
@@ -170,5 +158,9 @@ class TestStatTracker < Minitest::Test
     stat_tracker = StatTracker.new(locations)
 
     assert_equal '20132014', stat_tracker.worst_season('24')
+  end
+
+  def test_it_can_find_highest_total_score
+    assert_equal 11, @stat_tracker.highest_total_score
   end
 end
